EMGD package needs opengl in DISTRO_FEATURES
When opengl is not listed in DISTRO_FEATURES, the EMGD build fails as
seen below. The needed DRI, DRI2 and glx features are disabled in the
xserver-xorg when opengl is not listed in DISTRO_FEATURES.

|  * satisfy_dependencies_for: Cannot satisfy the following dependencies for packagegroup-core-x11-base:
|  *         xserver-xorg-extension-glx *     xserver-xorg-extension-dri *     xserver-xorg-extension-dri2 *
|  * opkg_install_cmd: Cannot install package packagegroup-core-x11-base.
| ERROR: Function failed: do_rootfs (log file is located at /srv/home/nitin/builds/build-minnow/tmp-eglibc/work/minnow-oe-linux/core-image-sato/1.0-r0/temp/log.do_rootfs.19450)
ERROR: Task 7 (/home/nitin/prj/openembedded-core.git/meta/recipes-sato/images/core-image-sato.bb, do_rootfs) failed with exit code '1'

This is not an issue when building with the poky distro as it lists
opengl in DISTRO_FEATURES. But when building with an empty DISTRO, this
fix is needed to avoid these build errors.

Signed-off-by: Nitin A Kamble <nitin.a.kamble@intel.com>
<dvhart: slight reordering to group emgd features>
Signed-off-by: Darren Hart <dvhart@linux.intel.com>