linux-yocto: Add bbappend for linux-yocto-3.10
This new bbappend adds support for the 3.10 linux-yocto kernel
repository and an updated BSP based on the standard/base branch with the
minnow-io feature used to encapsulate the GPIO platform drivers, which
are still out-of-tree. All the remaining drivers, emgd excepting of
course, are now backports from upstream. A new set of wifi features were
added to the 3.10 tree and are included in the bbappend as well.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>