minnow.conf: Prefer the 3.10 linux-yocto recipe
Update to the 3.10 kernel. This update bring us closer to a fully
upstream BSP. The 3.10 BSP uses the standard/base branch (all upstream
sources there) and adds the minnow-io feature as a selectable feature if
you want to build the GPIO LED and Button drivers which are not
upstreamable, these are enabled by default. A new ACPI described version
is in the works. This also adds wifi drivers to the default build.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>